MutableConvexHulls

Documentation for MutableConvexHulls.

MutableConvexHulls.MutableConvexHulls

MutableConvexHulls.HullNodeIterator

MutableConvexHulls.HullNodeIterator

MutableConvexHulls.MutableConvexHull

MutableConvexHulls.MutableLowerConvexHull

MutableConvexHulls.MutableUpperConvexHull

MutableConvexHulls.PointNodeIterator

MutableConvexHulls.PointNodeIterator

MutableConvexHulls.addpoint!

MutableConvexHulls.buildinreverse

MutableConvexHulls.firstpoint

MutableConvexHulls.insidehull

MutableConvexHulls.jarvismarch

MutableConvexHulls.jarvismarch!

MutableConvexHulls.lastpoint

MutableConvexHulls.lower_jarvismarch

MutableConvexHulls.lower_monotonechain

MutableConvexHulls.mergehulls!

MutableConvexHulls.mergepoints!

MutableConvexHulls.monotonechain

MutableConvexHulls.monotonechain!

MutableConvexHulls.removepoint!

MutableConvexHulls.upper_jarvismarch

MutableConvexHulls.upper_monotonechain

MutableConvexHulls.MutableConvexHulls — Module

MutableConvexHulls.jl provides ways to calculate and update vertex representations of planar convex polytopes (i.e. convex hulls). It is intended for use in situations when a convex hull must be updated iteratively with addition or removal of points.

See also MutableConvexHull, LowerMutableConvexHull, UpperMutableConvexHull, monotonechain, addpoint!, mergepoints!, removepoint!,

MutableConvexHulls.HullNodeIterator — Method

HullNodeIterator(hull [, rev])

Returns an iterator over the nodes in the linked list representing the convex hull.

If rev is true, the iterator will start at the tail of the list and advance toward the head. Otherwise, it will start at the head of the list and advance toward the tail.

MutableConvexHulls.HullNodeIterator — Method

HullNodeIterator(start [, rev])

Returns an iterator over the nodes in the linked list representing the convex hull, starting at the specified node start.

If rev is true, the iterator will advance toward the head of the list. Otherwise, it will advance toward the tail of the list.

MutableConvexHulls.MutableConvexHull — Method

h = MutableConvexHull{T}([, orientation, collinear, sortedby])

Initialize an empty MutableConvexHull with the provided attributes.

orientation specifies whether the points along the convex hull are ordered clockwise CW, or counterclockwise CCW, and defaults to CCW.

collinear specifies whether collinear points are allowed along the surface of the convex hull, and defaults to false.

sortedby specifies a function to apply to points prior to sorting, and defaults to identity (resulting in default sorting behavior).

See also: monotonechain, jarvismarch, addpoint!, mergepoints!, removepoint!

MutableConvexHulls.MutableLowerConvexHull — Method

h = MutableLowerConvexHull{T}([, orientation, collinear, sortedby])

Initialize an empty MutableLowerConvexHull with the provided attributes.

orientation specifies whether the points along the convex hull are ordered clockwise CW, or counterclockwise CCW, and defaults to CCW.

collinear specifies whether collinear points are allowed along the surface of the convex hull, and defaults to false.

sortedby specifies a function to apply to points prior to sorting, and defaults to identity (resulting in default sorting behavior).

See also: lower_monotonechain, lower_jarvismarch, addpoint!, mergepoints!, removepoint!

MutableConvexHulls.MutableUpperConvexHull — Method

h = MutableUpperConvexHull{T}([, orientation, collinear, sortedby])

Initialize an empty MutableUpperConvexHull with the provided attributes.

orientation specifies whether the points along the convex hull are ordered clockwise CW, or counterclockwise CCW, and defaults to CCW.

collinear specifies whether collinear points are allowed along the surface of the convex hull, and defaults to false.

sortedby specifies a function to apply to points prior to sorting, and defaults to identity (resulting in default sorting behavior).

See also: upper_monotonechain, upper_jarvismarch, addpoint!, mergepoints!, removepoint!

MutableConvexHulls.PointNodeIterator — Method

PointNodeIterator(hull [, rev])

Returns an iterator over the nodes in the linked list representing the points contained by the convex hull.

If rev is true, the iterator will start at the tail of the list and advance toward the head. Otherwise, it will start at the head of the list and advance toward the tail.

MutableConvexHulls.PointNodeIterator — Method

PointNodeIterator(start [, rev])

Returns an iterator over the nodes in the linked list representing the points contained by the convex hull, starting at the specified node start.

If rev is true, the iterator will advance toward the head of the list. Otherwise, it will advance toward the tail of the list.

MutableConvexHulls.addpoint! — Method

addpoint!(hull, point)

Add point to the list of points contained by the provided convex hull hull. If point lies outside the convex hull, the list of hull points will be updated accordingly.

See also: mergepoints!, removepoint!

MutableConvexHulls.buildinreverse — Method

buildinreverse(h) -> Bool

Return true if the monotonechain algorithm should start at the last point in h.points, and false otherwise.

MutableConvexHulls.firstpoint — Method

node = firstpoint(h)

Get the first point on h.points that should be considered for the monotone chain algorithm.

MutableConvexHulls.insidehull — Method

insidehull(data, hull) -> Bool

Return true if the data lies within the interior of hull and false otherwise.

MutableConvexHulls.jarvismarch! — Method

jarvismarch!(hull [, initedge, stop])

Determine the convex hull of the points contained in the provided hull.points using the Jarvis march algorithm. Each node in the list should contain a two-dimensional point, and the nodes are assumed to be sorted (e.g. by lowest "x" value and by lowest "y" in case of ties, though some other sorting methods may produce valid results).

MutableConvexHulls.jarvismarch — Method

h = jarvismarch(points [; orientation, collinear, sortedby])

Return the convex hull generated from the provided points.

orientation specifies whether the points along the convex hull are ordered clockwise CW, or counterclockwise CCW, and defaults to CCW.

collinear specifies whether collinear points are allowed along the surface of the convex hull, and defaults to false.

sortedby specifies a function to apply to points prior to sorting, and defaults to identity (resulting in default sorting behavior).

MutableConvexHulls.lastpoint — Method

node = lastpoint(h)

Get the last point on h.points that should be considered for the monotone chain algorithm.

MutableConvexHulls.lower_jarvismarch — Method

lh = lower_jarvismarch(points [; orientation, collinear, sortedby])

Return the lower convex hull generated from the provided points.

orientation specifies whether the points along the convex hull are ordered clockwise CW, or counterclockwise CCW, and defaults to CCW.

collinear specifies whether collinear points are allowed along the surface of the convex hull, and defaults to false.

sortedby specifies a function to apply to points prior to sorting, and defaults to identity (resulting in default sorting behavior).

MutableConvexHulls.lower_monotonechain — Method

lh = lower_monotonechain(points [; orientation, collinear, sortedby])

Return the lower convex hull generated from the provided points.

orientation specifies whether the points along the convex hull are ordered clockwise CW, or counterclockwise CCW, and defaults to CCW.

collinear specifies whether collinear points are allowed along the surface of the convex hull, and defaults to false.

sortedby specifies a function to apply to points prior to sorting, and defaults to identity (resulting in default sorting behavior).

MutableConvexHulls.mergehulls! — Method

mergehulls!(hull, otherhulls...)

Merge the points contained in otherhulls into hull. See Chan's algorithm for a similar approach.

MutableConvexHulls.mergepoints! — Method

mergepoints!(hull, points)

Add points to the list of points contained by the provided convex hull hull. If any of the points lie outside the convex hull, the list of hull points will be updated accordingly.

This function finds the convex hull of the points to be added before merging them with the hull. See Chan's algorithm for a similar idea.

See also: addpoint!, removepoint!

MutableConvexHulls.monotonechain! — Method

monotonechain!(hull [, start, stop])

Determine the convex hull of the points contained in the provided hull.points using the monotone chain algorithm. Each node in the list should contain a two-dimensional point, and the nodes are assumed to be sorted (e.g. by lowest "x" value and by lowest "y" in case of ties, though some other sorting methods may produce valid results).

start and stop should be nodes contained in hull.points. If start and/or stop are provided, the convex hull will only be updated on the inclusive interval between start and stop. This allows efficient removal of points from a convex hull.

MutableConvexHulls.monotonechain — Method

h = monotonechain(points [; orientation, collinear, sortedby])

Return the convex hull generated from the provided points.

orientation specifies whether the points along the convex hull are ordered clockwise CW, or counterclockwise CCW, and defaults to CCW.

collinear specifies whether collinear points are allowed along the surface of the convex hull, and defaults to false.

sortedby specifies a function to apply to points prior to sorting, and defaults to identity (resulting in default sorting behavior).

MutableConvexHulls.removepoint! — Method

removepoint!(hull, node)

Removes node from hull. If the node corresponds to a point on the convex hull, the list of hull points will be updated accordingly.

See also: addpoint!, mergepoints!

MutableConvexHulls.upper_jarvismarch — Method

uh = upper_jarvismarch(points [; orientation, collinear, sortedby])

Return the upper convex hull generated from the provided points.

orientation specifies whether the points along the convex hull are ordered clockwise CW, or counterclockwise CCW, and defaults to CCW.

collinear specifies whether collinear points are allowed along the surface of the convex hull, and defaults to false.

sortedby specifies a function to apply to points prior to sorting, and defaults to identity (resulting in default sorting behavior).

MutableConvexHulls.upper_monotonechain — Method

uh = upper_monotonechain(points [; orientation, collinear, sortedby])

Return the upper convex hull generated from the provided points.

orientation specifies whether the points along the convex hull are ordered clockwise CW, or counterclockwise CCW, and defaults to CCW.

collinear specifies whether collinear points are allowed along the surface of the convex hull, and defaults to false.

sortedby specifies a function to apply to points prior to sorting, and defaults to identity (resulting in default sorting behavior).